마스터 브랜치 hello world 로직변경

+++ Calculator2/src/Calculator2.java
@@ -2,8 +2,8 @@
 // Add 1 # add2 #add3 #add4 #5 new add 1 #6 new add2
 // B add
 public class Calculator2 {
-	public void helloworld(int code) {
-		
+	public void helloworld(float code) {
+		// 1, 2, 3, 4, 5
 	}
 
 }
